backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
gspread==5.12.0
google-auth==2.23.3
email-validator==2.0.0
pytest==7.4.3 